fix(ND): Fixed ND asterisk font #8462
Summary of Changes
A very important change: the PFD asterisk (*) font was not accurate, so I changed the ECAMFontRegular to match the reference pictures from issue #8343.

References
See issue #8343, good close-ups are hard to find, and sometimes they don't agree.

Additional context
The old asterisk font was (I think) from this font, but there is no evidence that it has ever been implemented in the A320neo. I checked as much as I could that the asterisk is not used anywhere else, and I only found the DCDU that has its own font.